fix(deploy): P0 deploy-unblocker — split flag boot, fix bundle picker
DGAI Netlify staging has been failing intermittently on bundle budget
since 997c2e32 (AnalysisHeroV17). Root cause: src/main.tsx:7 eagerly
imports `comparisonFlagBoot`, which pulled the @/flags dependency
chain into the entry chunk. Entry chunk has been at 49.75 KiB / 50
KiB (99.5% used) for ~5 days; every deploy is a coin-flip against
Netlify's gzip drift.

Latent bug: scripts/verify-bundle-budget.mjs:50-52 used Array.find
on readdirSync output, which is filesystem-order-dependent. On macOS
APFS it returned the small 9 KiB lucide companion (false-positive
PASS); on Netlify Linux ext4 hash-order it sometimes returned the
real 49.75 KiB entry. Local builds had been passing silently while
the real bundle was at the budget knife-edge.

Fix (narrow, P0 scope):
1. Split comparisonFlagBoot.ts:
   - Keep `bootAnalysisHeroCompareFromUrl()` lightweight — just the
     URL → localStorage write. No @/flags import. Stays eager-imported
     at main.tsx:7 so the URL → localStorage copy completes BEFORE
     StreamFlagsProvider's useState initialiser reads localStorage.
   - Move `window.__analysisHeroV17` dev-console diagnostics helper
     into a new sibling file `comparisonFlagDevHelper.ts` that owns
     the @/flags import. Exported as `installAnalysisHeroV17DevHelper()`.
2. main.tsx schedules a dynamic import of the dev helper inside a
   `requestIdleCallback` (with `setTimeout` fallback for browsers
   without rIC) after the initial Shell render. Dev helper is
   available in browser devtools but contributes zero weight to
   first-paint bundle.
3. Fix scripts/verify-bundle-budget.mjs picker: resolve the entry
   chunk via `dist/index.html`'s `<script type="module" src=...>`
   reference (the authoritative source). Fall back to the LARGEST
   `index-*.js` (size-ordered, not FS-order-dependent) if the HTML
   lookup misses.

Measurements (HEAD = 27153431, post-split):
- Entry chunk (Vite-reported gzip):  50.92 KB → 47.86 KB (Δ −3.06 KB)
- verify-bundle-budget.mjs gzipKB:    9.67 (companion picked, false pass)
                                   →  46.73 (real entry, true measurement)
- Budget margin: 0.25 KB (knife-edge) → 3.27 KB (healthy)
- New lazy chunk: comparisonFlagDevHelper-*.js, 1037 bytes raw

Verification:
- `npm run build:ci` clean (generate:validators + vite build +
  assert-plc-bundle + verify-bundle-budget all pass).
- `pnpm exec tsc --noEmit` clean.
- analysisHeroV17 targeted tests: 186/186 pass.
- No package.json / pnpm-lock.yaml / prompt / schema / backend changes.
- No router or boot-flow refactors; only the dev-helper portion is
  deferred. URL-parse path executes synchronously before first render
  exactly as before.

Branch base: origin/staging @ 27153431.
