+++ scripts/verify-bundle-budget.mjs
@@ -34,8 +34,21 @@ function getGzippedSize(filePath) {
 }
 
 /**
- * Find the main JavaScript bundle file
- * Vite generates files like: index-[hash].js
+ * Find the main JavaScript bundle file.
+ *
+ * Vite emits multiple `index-*.js` chunks per build (entry + lucide / debug-
+ * state companions). The real entry chunk is whichever filename is
+ * referenced from `dist/index.html` as `<script type="module" src=...>`.
+ * Resolving via the HTML is the only authoritative answer; using
+ * `readdirSync` + `Array.find` is filesystem-order-dependent (macOS APFS
+ * returns lex order; Linux ext4 returns hash order) and silently picks
+ * a non-entry companion on some platforms — masking budget breaches.
+ *
+ * Picker corrected 2026-05-13 (P0 deploy-unblocker). Previous behaviour
+ * was `files.find(f => f.startsWith('index-') && ...)` which on macOS
+ * was reporting the ~10 KB lucide companion instead of the ~50 KB
+ * real entry, so local `build:ci` had been passing while Netlify
+ * intermittently failed.
  */
 function findMainBundle(distPath) {
   if (!existsSync(distPath)) {
@@ -44,20 +57,36 @@ function findMainBundle(distPath) {
     return null
   }
 
-  const files = readdirSync(distPath)
-
-  // Look for index-*.js (main bundle)
-  const mainBundle = files.find(file =>
-    file.startsWith('index-') && file.endsWith('.js') && !file.endsWith('.map')
-  )
-
-  if (!mainBundle) {
-    console.error('[Bundle Budget] ❌ Main bundle (index-*.js) not found')
-    console.error('[Bundle Budget]    Available files:', files.join(', '))
-    return null
+  // Primary: resolve via dist/index.html's <script type="module" src=...>
+  const distRoot = join(distPath, '..')
+  const indexHtmlPath = join(distRoot, 'index.html')
+  if (existsSync(indexHtmlPath)) {
+    const html = readFileSync(indexHtmlPath, 'utf8')
+    // Match <script type="module" ... src="/assets/index-XXXX.js">
+    const match = html.match(/<script[^>]+type=["']module["'][^>]+src=["']\/?assets\/(index-[A-Za-z0-9_-]+\.js)["']/)
+    if (match && match[1]) {
+      const entryPath = join(distPath, match[1])
+      if (existsSync(entryPath)) {
+        return entryPath
+      }
+      console.error('[Bundle Budget] ⚠️  index.html references', match[1], 'but file missing from dist/assets; falling back to filesystem scan')
+    }
   }
 
-  return join(distPath, mainBundle)
+  // Fallback: pick the LARGEST `index-*.js` in dist/assets/. Deterministic
+  // (size order, not FS order). Mirrors Netlify behaviour better than
+  // Array.find but is still a fallback — primary path is the HTML lookup.
+  const files = readdirSync(distPath)
+    .filter(f => f.startsWith('index-') && f.endsWith('.js') && !f.endsWith('.map'))
+  if (files.length === 0) {
+    console.error('[Bundle Budget] ❌ Main bundle (index-*.js) not found')
+    console.error('[Bundle Budget]    Available files:', readdirSync(distPath).join(', '))
+    return null
+  }
+  const largest = files
+    .map(f => ({ name: f, size: statSync(join(distPath, f)).size }))
+    .sort((a, b) => b.size - a.size)[0]
+  return join(distPath, largest.name)
 }
 
 function main() {
