+++ src/components/results/analysisHeroV17/comparisonFlagBoot.ts
@@ -8,17 +8,16 @@
  * Per docs/brief-analysis-hero-v17-implementation.md §3 step 9. No per-render
  * URL parsing — the parse runs once at boot.
  *
- * This module also installs `window.__analysisHeroV17` as a dev-console
- * diagnostics helper. Type it in browser devtools to see flag state and
- * toggle helpers ("Why is my hero not showing?" diagnostics).
+ * The dev-console `window.__analysisHeroV17` diagnostics helper has been
+ * split into a separate module (`comparisonFlagDevHelper.ts`) loaded lazily
+ * from `main.tsx` after `createRoot`. This keeps the entry chunk free of
+ * the `@/flags` dependency chain — the URL-parse path runs synchronously
+ * before the first render (so `StreamFlagsProvider`'s `useState` initialiser
+ * sees the post-URL localStorage state), while the dev helper, which is
+ * only needed in browser devtools and not on the first paint, defers until
+ * after the user's content has rendered.
  */
 
-import {
-  isAnalysisHeroV17Enabled,
-  isAnalysisHeroCompareEnabled,
-} from '@/flags'
-
-const V17_STORAGE_KEY = 'feature.analysisHeroV17'
 const COMPARE_STORAGE_KEY = 'feature.analysisHeroCompare'
 
 export function bootAnalysisHeroCompareFromUrl(): void {
@@ -35,68 +34,5 @@ export function bootAnalysisHeroCompareFromUrl(): void {
     // localStorage / env-var state untouched.
   } catch {
     // SSR, tests, restricted localStorage — silently skip.
-  }
-
-  // Install a dev-console diagnostics helper. Always available (not just
-  // in DEV) so staging reviewers can self-diagnose by typing
-  // `__analysisHeroV17` in browser devtools without rebuilds.
-  try {
-    if (typeof window === 'undefined') return
-    const helper = {
-      get status() {
-        return {
-          analysisHeroV17: {
-            enabled: isAnalysisHeroV17Enabled(),
-            localStorage: readStorage(V17_STORAGE_KEY),
-          },
-          analysisHeroCompare: {
-            enabled: isAnalysisHeroCompareEnabled(),
-            localStorage: readStorage(COMPARE_STORAGE_KEY),
-          },
-        }
-      },
-      enable() {
-        try { localStorage.setItem(V17_STORAGE_KEY, '1') } catch {}
-        return helper.status
-      },
-      disable() {
-        try { localStorage.setItem(V17_STORAGE_KEY, '0') } catch {}
-        return helper.status
-      },
-      reset() {
-        try {
-          localStorage.removeItem(V17_STORAGE_KEY)
-          localStorage.removeItem(COMPARE_STORAGE_KEY)
-        } catch {}
-        return helper.status
-      },
-      enableCompare() {
-        try { localStorage.setItem(COMPARE_STORAGE_KEY, '1') } catch {}
-        return helper.status
-      },
-      disableCompare() {
-        try { localStorage.setItem(COMPARE_STORAGE_KEY, '0') } catch {}
-        return helper.status
-      },
-    }
-    // Attach to window for devtools access. Non-enumerable so it
-    // doesn't clutter standard window inspection.
-    Object.defineProperty(window, '__analysisHeroV17', {
-      value: helper,
-      writable: true,
-      configurable: true,
-      enumerable: false,
-    })
-  } catch {
-    // Silent — the helper is a convenience, not a load-bearing path.
-  }
-}
-
-function readStorage(key: string): string | null {
-  try {
-    if (typeof localStorage === 'undefined') return null
-    return localStorage.getItem(key)
-  } catch {
-    return null
   }
 }
